requirements.txt
aiohttp==3.8.6
aiomysql==0.2.0
aiopg==1.4.0
aioredis==2.0.1
aiosignal==1.3.1
alibabacloud-credentials==0.3.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-green20220302==1.1.0
alibabacloud-openapi-util==0.2.2
alibabacloud-tea==0.3.3
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.11
alibabacloud-tea-xml==0.0.2
annotated-types==0.6.0
anyio==3.7.1
async-lru==2.0.4
async-timeout==4.0.3
asyncache==0.3.1
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
databases==0.8.0
ecdsa==0.18.0
emoji==2.8.0
fastapi==0.104.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
html2text==2020.1.16
idna==3.4
jieba==0.42.1
Jinja2==3.1.2
langdetect==1.0.9
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
pip==23.3.1
prometheus-client==0.18.0
prometheus-fastapi-instrumentator==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
wheel==0.41.2
uvicorn==0.23.2
yarl==1.9.2uvicorn
databases==0.8.0